fix: Refactor carpetas de notifications
