+++ controllers/dashboard.controller.js
@@ -1,4 +1,4 @@
-const Notification = require('../models/notification.model');
+const Notification = require('../models/notifications/notification.model');
 
 exports.getDashboard = async (req, res) => {
     res.render('../views/dashboard.ejs', { title: 'Dashboard' });
