+++ views/notifications/notifications.ejs
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/7.0.1/css/all.min.css">
     <title>Notifications</title>
-    <link rel="stylesheet" href="/css/notifications.css">
+    <link rel="stylesheet" href="/css/notifications/notifications.css">
 </head>
 </head>
 <body>
@@ -76,7 +76,7 @@
     <div id="modal-root"></div>
 
     <div id="backdrop" class="backdrop"></div>
-    <script src="/js/editNotifications.js" defer></script>
+    <script src="/js/notifications/editNotifications.js" defer></script>
 </body>
 </html>
